requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.20
werkzeug==3.0.1

pymupdf==1.24.10
Pillow==10.4.0

faiss-cpu==1.8.0.post1
numpy==1.26.4
scikit-learn==1.5.1
rank-bm25==0.2.2

sentencepiece==0.2.0
langdetect==1.0.9
requests==2.32.3
starlette==0.38.2


# In-process CPU inference for GGUF models (Qwen2.5-7B quantized)
llama-cpp-python==0.3.16

# Optional: sentence-transformers for SBERT/BGE embeddings
sentence-transformers==3.0.1
